+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2024-09-25
+
+### Added
+
+- Added support for Personalise rules within `Mutually Exclusive Groups`.
+
 ## [1.1.0] - 2024-08-29
 
 ### Fixed
@@ -11,8 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix: Check for None values in `user_agent` and `ip_address` when sending impressions to VWO.
 
 ## [1.0.0] - 2024-06-20
+
 ### Added
+
 - First release of VWO Feature Management and Experimentation capabilities
+
   ```python
   from vwo import init
 
@@ -48,7 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
           return client_db.get(f"{key}_{user_id}")
 
       def set(self, value: dict):
-          key = f"{value.get('featureKey')}_{value.get('user')}"
+          key = f"{value.get('featureKey')}_{value.get('userId')}"
           client_db[key] = {
               'rolloutKey': value.get('rolloutKey'),
               'rolloutVariationId': value.get('rolloutVariationId'),
@@ -58,7 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
               'experimentId': value.get('experimentId'),
           }
           return True
-        
+
   options = {
     'sdk_key': '32-alpha-numeric-sdk-key', # SDK Key
     'account_id': '123456', # VWO Account ID
@@ -94,4 +103,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   vwo_client = init(options)
   ```
-